// src/types/cityProps.ts
export interface cityProps {
  code: string
  location: Location[]
  refer?: Refer
}

export interface Refer {
  sources?: string[]
  license?: string[]
}

export interface Location {
  name: string
  id: string
  lat: string
  lon: string
  adm2: string
  adm1: string
  country: string
  tz: string
  utcOffset: string
  isDst: string
  type: string
  rank: string
  fxLink: string
}